requirement.txt
beautifulsoup4==4.5.3
bs4==0.0.1
catfish==1.4.2
chardet==2.3.0
command-not-found==0.3
defer==1.0.6
Django==1.10.5
django-datetime-widget==0.9.3
django-swingtime==0.7.1
django-webtest==1.9.2
get==0.0.0
html5lib==0.999
language-selector==0.1
lightdm-gtk-greeter-settings==1.2.1
lxml==3.5.0
menulibre==2.1.3
mugshot==0.3.1
nltk==3.2.2
onboard==1.2.0
pexpect==4.0.1
Pillow==3.1.2
post==0.0.0
psutil==3.4.2
ptyprocess==0.5
public==0.0.0
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyPDF2==1.26.0
python-apt==1.1.0b1
python-dateutil==2.6.0
python-debian==0.1.27
python-systemd==231
pytz==2016.10
pyxdg==0.25
query-string==0.0.0
reportlab==3.3.0
request==0.0.0
requests==2.9.1
sessioninstaller==0.0.0
setupfiles==0.0.0
six==1.10.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
waitress==1.0.2
WebOb==1.7.2
WebTest==2.0.27
xkit==0.0.0


